working version of conversion to v3
