+++ config.js
@@ -1,0 +1,337 @@
+// Brightness
+exports.CHOICES_BRIGHTNESS = [
+    { id: '0',  label: '3%',   cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x08,0x5D,0x5A]) },
+    { id: '1',  label: '5%',   cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x0D,0x62,0x5A]) },
+    { id: '2',  label: '8%',   cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x14,0x69,0x5A]) },
+    { id: '3',  label: '10%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x1A,0x6F,0x5A]) },
+    { id: '4',  label: '15%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x27,0x7C,0x5A]) },
+    { id: '5',  label: '20%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x33,0x88,0x5A]) },
+    { id: '6',  label: '25%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x40,0x95,0x5A]) },
+    { id: '7',  label: '30%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x4D,0xA2,0x5A]) },
+    { id: '8',  label: '35%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x5A,0xAF,0x5A]) },
+    { id: '9',  label: '40%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x66,0xBB,0x5A]) },
+    { id: '10', label: '45%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x73,0xC8,0x5A]) },
+    { id: '11', label: '50%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x80,0xD5,0x5A]) },
+    { id: '12', label: '55%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x8D,0xE2,0x5A]) },
+    { id: '13', label: '60%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x9A,0xEF,0x5A]) },
+    { id: '14', label: '65%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xA6,0xFB,0x5A]) },
+    { id: '15', label: '70%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xB3,0x08,0x5B]) },
+    { id: '16', label: '75%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xC0,0x15,0x5B]) },
+    { id: '17', label: '80%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xCD,0x22,0x5B]) },
+    { id: '18', label: '85%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xDA,0x2F,0x5B]) },
+    { id: '19', label: '90%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xE6,0x3B,0x5B]) },
+    { id: '20', label: '95%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xF3,0x48,0x5B]) },
+    { id: '21', label: '100%', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xFF,0x54,0x5B]) }
+];
+
+// Brightness
+exports.CHOICES_BRIGHTNESS_VX1000 = [
+    { id: '0',  label: '0%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x00,0x55,0x5A]) },
+    { id: '1',  label: '10%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x19,0x6E,0x5A]) },
+    { id: '2',  label: '20%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x33,0x88,0x5A]) },
+    { id: '3',  label: '30%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x4C,0xA1,0x5A]) },
+    { id: '4',  label: '40%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x66,0xBB,0x5A]) },
+    { id: '5',  label: '50%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x7F,0xD4,0x5A]) },
+    { id: '6',  label: '60%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0x99,0xEE,0x5A]) },
+    { id: '7',  label: '70%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xB2,0x07,0x5B]) },
+    { id: '8',  label: '80%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xCC,0x21,0x5B]) },
+    { id: '9',  label: '90%',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xE5,0x3A,0x5B]) },
+    { id: '10', label: '100%', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x01,0xFF,0xFF,0xFF,0x01,0x00,0x01,0x00,0x00,0x02,0x01,0x00,0xFF,0x5A,0x5B]) }
+];
+
+// Test Patterns
+exports.CHOICES_TESTPATTERNS = [
+    { id: '0', label: 'Red',        cmd: Buffer.from([0x55,0xAA,0x00,0x80,0xFE,0x00,0x01,0x00,0xFF,0xFF,0x01,0x00,0x01,0x01,0x00,0x02,0x01,0x00,0x02,0xDA,0x58]) },
+    { id: '1', label: 'Green',      cmd: Buffer.from([0x55,0xAA,0x00,0x80,0xFE,0x00,0x01,0x00,0xFF,0xFF,0x01,0x00,0x01,0x01,0x00,0x02,0x01,0x00,0x03,0xDB,0x58]) },
+    { id: '2', label: 'Blue',       cmd: Buffer.from([0x55,0xAA,0x00,0x80,0xFE,0x00,0x01,0x00,0xFF,0xFF,0x01,0x00,0x01,0x01,0x00,0x02,0x01,0x00,0x04,0xDC,0x58]) },
+    { id: '3', label: 'White',      cmd: Buffer.from([0x55,0xAA,0x00,0x80,0xFE,0x00,0x01,0x00,0xFF,0xFF,0x01,0x00,0x01,0x01,0x00,0x02,0x01,0x00,0x05,0xDD,0x58]) },
+    { id: '4', label: 'Horizontal', cmd: Buffer.from([0x55,0xAA,0x00,0x80,0xFE,0x00,0x01,0x00,0xFF,0xFF,0x01,0x00,0x01,0x01,0x00,0x02,0x01,0x00,0x06,0xDE,0x58]) },
+    { id: '5', label: 'Vertical',   cmd: Buffer.from([0x55,0xAA,0x00,0x80,0xFE,0x00,0x01,0x00,0xFF,0xFF,0x01,0x00,0x01,0x01,0x00,0x02,0x01,0x00,0x07,0xDF,0x58]) },
+    { id: '6', label: 'Diagonal',   cmd: Buffer.from([0x55,0xAA,0x00,0x80,0xFE,0x00,0x01,0x00,0xFF,0xFF,0x01,0x00,0x01,0x01,0x00,0x02,0x01,0x00,0x08,0xE0,0x58]) },
+    { id: '7', label: 'Gray-Scale', cmd: Buffer.from([0x55,0xAA,0x00,0x80,0xFE,0x00,0x01,0x00,0xFF,0xFF,0x01,0x00,0x01,0x01,0x00,0x02,0x01,0x00,0x09,0xE1,0x58]) },
+    { id: '8', label: 'Aging-All',  cmd: Buffer.from([0x55,0xAA,0x00,0x80,0xFE,0x00,0x01,0x00,0xFF,0xFF,0x01,0x00,0x01,0x01,0x00,0x02,0x01,0x00,0x0A,0xE2,0x58]) }
+];
+
+// VX4S Display Modes
+exports.CHOICES_DISPLAYMODE = [
+    { id: '0', label: 'Normal', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x50,0x00,0x20,0x02,0x01,0x00,0x00,0xC7,0x56]) },
+    { id: '1', label: 'Freeze', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x50,0x00,0x20,0x02,0x01,0x00,0x01,0xC8,0x56]) },
+    { id: '2', label: 'Black',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x50,0x00,0x20,0x02,0x01,0x00,0x02,0xC9,0x56]) }
+];
+
+// VX6S Display Modes
+exports.CHOICES_DISPLAYMODE_VX6S = [
+    { id: '0', label: 'Normal', cmd: Buffer.from([0x55,0xAA,0x00,0x38,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x04,0x00,0x00,0x13,0x01,0x00,0x03,0xA7,0x56]) },
+    { id: '1', label: 'Freeze', cmd: Buffer.from([0x55,0xAA,0x00,0x35,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x04,0x00,0x00,0x13,0x01,0x00,0x03,0xA5,0x56]) },
+    { id: '2', label: 'Black',  cmd: Buffer.from([0x55,0xAA,0x00,0x37,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x04,0x00,0x00,0x13,0x01,0x00,0x03,0xA2,0x56]) }
+];	
+
+// VX6S Working Modes
+exports.CHOICES_WORKINGMODE_VX6S = [
+    { id: '0', label: 'Direct',   cmd: Buffer.from([0x55,0xAA,0x00,0xEC,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2C,0x00,0x00,0x13,0x01,0x00,0x00,0x80,0x57]) },
+    { id: '1', label: 'Switcher', cmd: Buffer.from([0x55,0xAA,0x00,0x1E,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2C,0x00,0x00,0x13,0x01,0x00,0x01,0xB3,0x56]) }
+];					
+
+// NovaProUHD Jr & VX1000 Display Modes
+exports.CHOICES_DISPLAYMODE_VX1000 = [
+    { id: '0', label: 'Normal', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x04,0x00,0x00,0x13,0x02,0x00,0x03,0x00,0x70,0x56]) },
+    { id: '1', label: 'Freeze', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x04,0x00,0x00,0x13,0x02,0x00,0x04,0x00,0x71,0x56]) },
+    { id: '2', label: 'Black',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x04,0x00,0x00,0x13,0x02,0x00,0x05,0x00,0x72,0x56]) }
+];
+
+// VX6S Working Modes
+exports.CHOICES_WORKINGMODE_VX6S = [
+    { id: '0', label: 'Direct',   cmd: Buffer.from([0x55,0xAA,0x00,0xEC,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2C,0x00,0x00,0x13,0x01,0x00,0x00,0x80,0x57]) },
+    { id: '1', label: 'Switcher', cmd: Buffer.from([0x55,0xAA,0x00,0x1E,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2C,0x00,0x00,0x13,0x01,0x00,0x01,0xB3,0x56]) }
+];					
+
+// VX6S Working Modes
+exports.CHOICES_WORKINGMODE_J6 = [
+    { id: '0', label: 'Splicer Mode',   cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2C,0x00,0x00,0x13,0x01,0x00,0x00,0x94,0x56]) },
+    { id: '1', label: 'Switcher Mode', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2C,0x00,0x00,0x13,0x01,0x00,0x01,0x95,0x56]) }
+];
+// MCTRL4K Inputs
+exports.CHOICES_INPUTS_MCTRL4K = [
+    { id: '0', label: 'DVI',          cmd: Buffer.from([0x55,0xAA,0x00,0x3E,0xFE,0xFF,0x00,0x00,0x00,0x00,0x01,0x00,0x23,0x00,0x00,0x02,0x01,0x00,0x61,0x18,0x58]) },
+    { id: '1', label: 'HDMI',         cmd: Buffer.from([0x55,0xAA,0x00,0x8A,0xFE,0xFF,0x00,0x00,0x00,0x00,0x01,0x00,0x23,0x00,0x00,0x02,0x01,0x00,0x05,0x08,0x58]) },
+    { id: '2', label: 'Display Port', cmd: Buffer.from([0x55,0xAA,0x00,0x9D,0xFE,0xFF,0x00,0x00,0x00,0x00,0x01,0x00,0x23,0x00,0x00,0x02,0x01,0x00,0x5F,0x75,0x58]) }
+];
+
+// VX4S Inputs
+exports.CHOICES_INPUTS_VX4S = [
+    { id: '0', label: 'DVI',          cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2D,0x00,0x20,0x02,0x01,0x00,0x10,0xB4,0x56]) },
+    { id: '1', label: 'HDMI',         cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2D,0x00,0x20,0x02,0x01,0x00,0xA0,0x44,0x57]) },
+    { id: '2', label: 'VGA 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2D,0x00,0x20,0x02,0x01,0x00,0x01,0xA5,0x56]) },
+    { id: '3', label: 'VGA 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2D,0x00,0x20,0x02,0x01,0x00,0x02,0xA6,0x56]) },
+    { id: '4', label: 'CVBS 1',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2D,0x00,0x20,0x02,0x01,0x00,0x71,0x15,0x57]) },
+    { id: '5', label: 'CVBS 2',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2D,0x00,0x20,0x02,0x01,0x00,0x72,0x16,0x57]) },
+    { id: '6', label: 'SDI',          cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2D,0x00,0x20,0x02,0x01,0x00,0x40,0xE4,0x56]) },
+    { id: '7', label: 'Display Port', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x2D,0x00,0x20,0x02,0x01,0x00,0x90,0x34,0x57]) }
+];
+
+// VX6S Inputs
+exports.CHOICES_INPUTS_VX6S = [
+    { id: '0', label: '[Window 1] HDMI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x88,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x00,0x00,0x11,0x17,0x57]) },
+    { id: '1', label: '[Window 1] HDMI 2',        cmd: Buffer.from([0x55,0xAA,0x00,0xA8,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x01,0x00,0x12,0x39,0x57]) },
+    { id: '2', label: '[Window 1] SDI 1',         cmd: Buffer.from([0x55,0xAA,0x00,0xC4,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x02,0x00,0x31,0x75,0x57]) },
+    { id: '3', label: '[Window 1] SDI 2',         cmd: Buffer.from([0x55,0xAA,0x00,0xD4,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x03,0x00,0x32,0x87,0x57]) },
+    { id: '4', label: '[Window 1] DVI 1',         cmd: Buffer.from([0x55,0xAA,0x00,0xD6,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x04,0x00,0x01,0x59,0x57]) },
+    { id: '5', label: '[Window 1] DVI 2',         cmd: Buffer.from([0x55,0xAA,0x00,0xD7,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x06,0x00,0x02,0x5D,0x57]) },
+    { id: '6', label: '[Window 1] USB',           cmd: Buffer.from([0x55,0xAA,0x00,0xD9,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x07,0x00,0xA0,0xFE,0x57]) },
+    { id: '7', label: '[Window 2] HDMI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x47,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x00,0x01,0x11,0x07,0x57]) },
+    { id: '9', label: '[Window 2] HDMI 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x55,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x01,0x01,0x12,0x17,0x57]) },
+    { id: '10', label: '[Window 2] SDI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x56,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x02,0x01,0x31,0x38,0x57]) },
+    { id: '11', label: '[Window 2] SDI 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x58,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x03,0x01,0x32,0x3C,0x57]) },
+    { id: '12', label: '[Window 2] DVI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x59,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x04,0x01,0x01,0x0D,0x57]) },
+    { id: '13', label: '[Window 2] DVI 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x5A,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x06,0x01,0x02,0x11,0x57]) },
+    { id: '14', label: '[Window 2] USB',          cmd: Buffer.from([0x55,0xAA,0x00,0x5C,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x07,0x01,0xA0,0xB2,0x57]) },
+    { id: '15', label: '[Window 3] HDMI 1',       cmd: Buffer.from([0x55,0xAA,0x00,0xA2,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x00,0x02,0x11,0x93,0x57]) },
+    { id: '16', label: '[Window 3] HDMI 2',       cmd: Buffer.from([0x55,0xAA,0x00,0xA3,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x01,0x02,0x12,0x96,0x57]) },
+    { id: '17', label: '[Window 3] SDI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0xA5,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x02,0x02,0x31,0xB8,0x57]) },
+    { id: '18', label: '[Window 3] SDI 2',        cmd: Buffer.from([0x55,0xAA,0x00,0xA6,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x03,0x02,0x32,0xBB,0x57]) },
+    { id: '19', label: '[Window 3] DVI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0xA7,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x04,0x02,0x01,0x8C,0x57]) },
+    { id: '20', label: '[Window 3] DVI 2',        cmd: Buffer.from([0x55,0xAA,0x00,0xA9,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x06,0x02,0x02,0x91,0x57]) },
+    { id: '21', label: '[Window 3] USB',          cmd: Buffer.from([0x55,0xAA,0x00,0xAA,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x07,0x02,0xA0,0x31,0x58]) }
+];
+
+// NovaProHD Inputs
+exports.CHOICES_INPUTS_NOVAPROHD = [
+    { id: '0', label: 'SDI',          cmd: Buffer.from([0x55,0xAA,0x00,0x2B,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x22,0x00,0x20,0x02,0x01,0x00,0x1A,0xDE,0x56]) },
+    { id: '1', label: 'DVI',          cmd: Buffer.from([0x55,0xAA,0x00,0x34,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x22,0x00,0x20,0x02,0x01,0x00,0x1C,0xE9,0x56]) },
+    { id: '2', label: 'HDMI',         cmd: Buffer.from([0x55,0xAA,0x00,0x3F,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x22,0x00,0x20,0x02,0x01,0x00,0x1B,0xF3,0x56]) },
+    { id: '3', label: 'Display Port', cmd: Buffer.from([0x55,0xAA,0x00,0x51,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x22,0x00,0x20,0x02,0x01,0x00,0x1E,0x08,0x57]) },
+    { id: '4', label: 'VGA',          cmd: Buffer.from([0x55,0xAA,0x00,0x48,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x22,0x00,0x20,0x02,0x01,0x00,0x17,0xF8,0x56]) },
+    { id: '5', label: 'CVBS',         cmd: Buffer.from([0x55,0xAA,0x00,0x5A,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x22,0x00,0x20,0x02,0x01,0x00,0x02,0xF5,0x56]) }
+];
+
+// NovaProHD Inputs
+exports.CHOICES_INPUTS_NOVAPROUHDJR = [
+    { id: '0', label: 'SDI',          cmd: Buffer.from([0x55,0xAA,0x00,0x2B,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x22,0x00,0x20,0x02,0x01,0x00,0x1A,0xDE,0x56]) },
+    { id: '1', label: 'DVI',          cmd: Buffer.from([0x55,0xAA,0x00,0x34,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x22,0x00,0x20,0x02,0x01,0x00,0x1C,0xE9,0x56]) },
+    { id: '2', label: 'HDMI',         cmd: Buffer.from([0x55,0xAA,0x00,0x3F,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x22,0x00,0x20,0x02,0x01,0x00,0x1B,0xF3,0x56]) },
+    { id: '3', label: 'Display Port', cmd: Buffer.from([0x55,0xAA,0x00,0x51,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x22,0x00,0x20,0x02,0x01,0x00,0x1E,0x08,0x57]) }
+];
+
+// VX1000 Inputs
+exports.CHOICES_INPUTS_VX1000 = [
+    { id: '0', label: 'Layer 1-HDMI 1',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x00,0x00,0x00,0x7E,0x56]) },
+    { id: '1', label: 'Layer 1-HDMI 2',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x01,0x00,0x00,0x7F,0x56]) },
+    { id: '2', label: 'Layer 1-DVI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x02,0x00,0x00,0x80,0x56]) },
+    { id: '3', label: 'Layer 1-DVI 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x03,0x00,0x00,0x81,0x56]) },
+    { id: '4', label: 'Layer 1-SDI',          cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x04,0x00,0x00,0x82,0x56]) },
+    { id: '5', label: 'Layer 1-OPT 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x05,0x00,0x00,0x83,0x56]) },
+    { id: '6', label: 'Layer 1-OPT 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x06,0x00,0x00,0x84,0x56]) },
+    
+    { id: '7', label: 'Layer 2-HDMI 1',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x00,0x00,0x00,0xaE,0x56]) },
+    { id: '8', label: 'Layer 2-HDMI 2',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x01,0x00,0x00,0xaF,0x56]) },
+    { id: '9', label: 'Layer 2-DVI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x02,0x00,0x00,0xb0,0x56]) },
+    { id: '10', label: 'Layer 2-DVI 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x03,0x00,0x00,0xb1,0x56]) },
+    { id: '11', label: 'Layer 2-SDI',          cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x04,0x00,0x00,0xb2,0x56]) },
+    { id: '12', label: 'Layer 2-OPT 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x05,0x00,0x00,0xb3,0x56]) },
+    { id: '13', label: 'Layer 2-OPT 2',        cmd:Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x06,0x00,0x00,0xb4,0x56]) },
+    { id: '14', label: 'Layer 3-HDMI 1',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x00,0x00,0x00,0xDE,0x56]) },
+    { id: '15', label: 'Layer 3-HDMI 2',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x01,0x00,0x00,0xDF,0x56]) },
+    { id: '16', label: 'Layer 3-DVI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x02,0x00,0x00,0xE0,0x56]) },
+    { id: '17', label: 'Layer 3-DVI 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x03,0x00,0x00,0XE1,0x56]) },
+    { id: '18', label: 'Layer 3-SDI',          cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x04,0x00,0x00,0xE2,0x56]) },
+    { id: '19', label: 'Layer 3-OPT 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x05,0x00,0x00,0xE3,0x56]) },
+    { id: '20', label: 'Layer 3-OPT 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x06,0x00,0x00,0xE4,0x56]) }
+    
+];
+
+// VX600 Inputs
+exports.CHOICES_INPUTS_VX600 = [
+    { id: '0', label: 'Layer 1-HDMI 1',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x00,0x00,0x00,0x7E,0x56]) },
+    { id: '1', label: 'Layer 1-HDMI 2',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x01,0x00,0x00,0x7F,0x56]) },
+    { id: '2', label: 'Layer 1-DVI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x02,0x00,0x00,0x80,0x56]) },
+    { id: '3', label: 'Layer 1-SDI',          cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x04,0x00,0x00,0x82,0x56]) },
+    { id: '4', label: 'Layer 1-OPT 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x05,0x00,0x00,0x83,0x56]) },
+    { id: '5', label: 'Layer 1-OPT 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x06,0x00,0x00,0x84,0x56]) },
+    { id: '6', label: 'Layer 1- Mosaic',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x12,0x00,0x02,0x13,0x03,0x00,0x07,0x00,0x00,0x85,0x56]) },
+    { id: '7', label: 'Layer 2-HDMI 1',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x00,0x00,0x00,0xaE,0x56]) },
+    { id: '8', label: 'Layer 2-HDMI 2',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x01,0x00,0x00,0xaF,0x56]) },
+    { id: '9', label: 'Layer 2-DVI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x02,0x00,0x00,0xb0,0x56]) },
+    { id: '10', label: 'Layer 2-SDI',          cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x04,0x00,0x00,0xb2,0x56]) },
+    { id: '11', label: 'Layer 2-OPT 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x05,0x00,0x00,0xb3,0x56]) },
+    { id: '12', label: 'Layer 2-OPT 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x06,0x00,0x00,0xb4,0x56]) },
+    { id: '13', label: 'Layer 2-Mosaic',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x42,0x00,0x02,0x13,0x03,0x00,0x07,0x00,0x00,0xb5,0x56]) },
+    { id: '14', label: 'Layer 3-HDMI 1',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x00,0x00,0x00,0xDE,0x56]) },
+    { id: '15', label: 'Layer 3-HDMI 2',       cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x01,0x00,0x00,0xDF,0x56]) },
+    { id: '16', label: 'Layer 3-DVI 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x02,0x00,0x00,0xE0,0x56]) },
+    { id: '17', label: 'Layer 3-SDI',          cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x04,0x00,0x00,0xE2,0x56]) },
+    { id: '18', label: 'Layer 3-OPT 1',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x05,0x00,0x00,0xE3,0x56]) },
+    { id: '19', label: 'Layer 3-OPT 2',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x06,0x00,0x00,0xE4,0x56]) },
+    { id: '20', label: 'Layer 3-Mosaic',        cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x72,0x00,0x02,0x13,0x03,0x00,0x07,0x00,0x00,0XE5,0x56]) }
+];
+
+// PIP on off NovaPro HD & VX4S
+exports.CHOICES_PIP_ONOFF = [
+    { id: '0', label: 'Off', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x00,0x00,0x00,0x00,0x01,0x00,0x30,0x00,0x20,0x02,0x01,0x00,0x00,0xA6,0x57]) },
+    { id: '1', label: 'On',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x00,0x00,0x00,0x00,0x01,0x00,0x30,0x00,0x20,0x02,0x01,0x00,0x01,0xA7,0x57]) }
+];
+
+exports.CHOICES_PRESETS_VX4S = [
+    { id: '0',  label: 'Preset 1',  cmd: Buffer.from([0x55,0xAA,0x00,0x2E,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x70,0x00,0x20,0x02,0x01,0x00,0x00,0x15,0x57]) },
+    { id: '1',  label: 'Preset 2',  cmd: Buffer.from([0x55,0xAA,0x00,0x2E,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x70,0x00,0x20,0x02,0x01,0x00,0x01,0x16,0x57]) },
+    { id: '2',  label: 'Preset 3',  cmd: Buffer.from([0x55,0xAA,0x00,0x2E,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x70,0x00,0x20,0x02,0x01,0x00,0x02,0x17,0x57]) },
+    { id: '3',  label: 'Preset 4',  cmd: Buffer.from([0x55,0xAA,0x00,0x2E,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x70,0x00,0x20,0x02,0x01,0x00,0x03,0x18,0x57]) },
+    { id: '4',  label: 'Preset 5',  cmd: Buffer.from([0x55,0xAA,0x00,0x2E,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x70,0x00,0x20,0x02,0x01,0x00,0x04,0x19,0x57]) },
+    { id: '5',  label: 'Preset 6',  cmd: Buffer.from([0x55,0xAA,0x00,0x95,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x70,0x00,0x20,0x02,0x01,0x00,0x05,0x81,0x57]) },
+    { id: '6',  label: 'Preset 7',  cmd: Buffer.from([0x55,0xAA,0x00,0x95,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x70,0x00,0x20,0x02,0x01,0x00,0x06,0x82,0x57]) },
+    { id: '7',  label: 'Preset 8',  cmd: Buffer.from([0x55,0xAA,0x00,0x95,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x70,0x00,0x20,0x02,0x01,0x00,0x07,0x83,0x57]) },
+    { id: '8',  label: 'Preset 9',  cmd: Buffer.from([0x55,0xAA,0x00,0x95,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x70,0x00,0x20,0x02,0x01,0x00,0x08,0x84,0x57]) },
+    { id: '9',  label: 'Preset 10', cmd: Buffer.from([0x55,0xAA,0x00,0x95,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x70,0x00,0x20,0x02,0x01,0x00,0x09,0x85,0x57]) }
+];
+
+// VX6S Presets
+exports.CHOICES_PRESETS_VX6S = [
+    { id: '0',  label: 'Preset 1',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x00,0x05,0x16]) },
+    { id: '1',  label: 'Preset 2',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x01,0x06,0x16]) },
+    { id: '2',  label: 'Preset 3',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x02,0x07,0x56]) },
+    { id: '3',  label: 'Preset 4',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x03,0x08,0x16]) },
+    { id: '4',  label: 'Preset 5',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x04,0x09,0x16]) },
+    { id: '5',  label: 'Preset 6',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x05,0x0A,0x16]) },
+    { id: '6',  label: 'Preset 7',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x06,0x0B,0x56]) },
+    { id: '7',  label: 'Preset 8',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x07,0x0C,0x56]) },
+    { id: '8',  label: 'Preset 9',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x08,0x0D,0x56]) },
+    { id: '9',  label: 'Preset 10', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x09,0x0E,0x56]) },
+    { id: '10', label: 'Preset 11', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x0A,0x0F,0x56]) },
+    { id: '11', label: 'Preset 12', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x0B,0xC6,0x56]) },
+    { id: '12', label: 'Preset 13', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x0C,0xC7,0x56]) },
+    { id: '13', label: 'Preset 14', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x0D,0xC8,0x56]) },
+    { id: '14', label: 'Preset 15', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x0E,0xC9,0x56]) },
+    { id: '15', label: 'Preset 16', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x0F,0xCA,0x56]) }
+];
+
+// NovaProUHDJr Presets
+exports.CHOICES_PRESETS_NovaProUHDJr = [
+    { id: '0',  label: 'Preset 1',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x00,0x05,0x16]) },
+    { id: '1',  label: 'Preset 2',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x01,0x06,0x16]) },
+    { id: '2',  label: 'Preset 3',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x02,0x07,0x56]) },
+    { id: '3',  label: 'Preset 4',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x03,0x08,0x16]) },
+    { id: '4',  label: 'Preset 5',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x04,0x09,0x16]) },
+    { id: '5',  label: 'Preset 6',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x05,0x0A,0x16]) },
+    { id: '6',  label: 'Preset 7',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x06,0x0B,0x56]) },
+    { id: '7',  label: 'Preset 8',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x07,0x0C,0x56]) },
+    { id: '8',  label: 'Preset 9',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x08,0x0D,0x56]) },
+    { id: '9',  label: 'Preset 10', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x09,0x0E,0x56]) }
+];
+
+// VX1000 Presets
+exports.CHOICES_PRESETS_VX1000 = [
+    
+    { id: '0',  label: 'Preset 1',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x00,0xBA,0x56]) },
+    { id: '1',  label: 'Preset 2',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x01,0xBB,0x56]) },
+    { id: '2',  label: 'Preset 3',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x02,0xBC,0x56]) },
+    { id: '3',  label: 'Preset 4',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x03,0xBD,0x56]) },
+    { id: '4',  label: 'Preset 5',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x04,0xBE,0x56]) },
+    { id: '5',  label: 'Preset 6',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x05,0xBF,0x56]) },
+    { id: '6',  label: 'Preset 7',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x06,0xC0,0x56]) },
+    { id: '7',  label: 'Preset 8',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x07,0xC1,0x56]) },
+    { id: '8',  label: 'Preset 9',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x08,0xC2,0x56]) },
+    { id: '9',  label: 'Preset 10', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0x00,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x01,0x51,0x13,0x01,0x00,0x09,0xC3,0x56]) }
+];
+
+
+//Scaling Options - VX4S, VX6S, NovaProHD only
+exports.CHOICES_SCALING = [
+    { id: '0', label: 'Disable', cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x00,0x20,0x02,0x01,0x00,0x00,0x76,0x57]) },
+    { id: '1', label: 'Custom',  cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x00,0x20,0x02,0x01,0x00,0x01,0x77,0x57]) },
+    { id: '2', label: 'Auto',    cmd: Buffer.from([0x55,0xAA,0x00,0x00,0xFE,0xFF,0x00,0x00,0x00,0x00,0x01,0x00,0x00,0x00,0x20,0x02,0x01,0x00,0x02,0x78,0x57]) }
+];
+
+exports.CHOICES_LAYERS_VX1000 = [
+    { id: '0', label: 'Main Layer', cmd: Buffer.from([0x00]) },
+    { id: '1', label: 'Layer 1', cmd: Buffer.from([0x01]) },
+    { id: '2', label: 'Layer 2', cmd: Buffer.from([0x02]) }
+];
+
+exports.CHOICES_CARDNO_VX1000 = [
+    { id: '0', label: 'HDMI1', cmd: Buffer.from([0x00]) },
+    { id: '1', label: 'HDMI2', cmd: Buffer.from([0x01]) },
+    { id: '2', label: 'DVI1', cmd: Buffer.from([0x02]) },
+    { id: '3', label: 'DVI2', cmd: Buffer.from([0x03]) },
+    { id: '4', label: 'SDI', cmd: Buffer.from([0x04]) },
+    { id: '5', label: 'OPT-1', cmd: Buffer.from([0x05]) },
+    { id: '6', label: 'OPT-2', cmd: Buffer.from([0x06]) },
+    { id: '7', label: 'MOSAIC', cmd: Buffer.from([0x07]) }
+];
+
+exports.CHOICES_LAYERPRIORITY_VX1000 = [
+    { id: '0', label: 'Layer Priority 1', cmd: Buffer.from([0x00]) },
+    { id: '1', label: 'Layer Priority 2', cmd: Buffer.from([0x01]) },
+    { id: '2', label: 'Layer Priority 3', cmd: Buffer.from([0x02]) },
+];
+
+exports.CHOICES_CONNECTORCODE_VX1000 = [
+    { id: '0', label: 'HDMI1', cmd: Buffer.from([0x61]) },
+    { id: '1', label: 'HDMI2', cmd: Buffer.from([0x62]) },
+    { id: '2', label: 'DVI1', cmd: Buffer.from([0x69]) },
+    { id: '3', label: 'DVI2', cmd: Buffer.from([0x6A]) },
+    { id: '4', label: 'SDI', cmd: Buffer.from([0x30]) },
+    { id: '5', label: 'OPT-1', cmd: Buffer.from([0xE9]) },
+    { id: '6', label: 'OPT-2', cmd: Buffer.from([0xEA]) },
+    { id: '7', label: 'MOSAIC', cmd: Buffer.from([0xE0]) }
+];
+
+exports.CONFIG_MODEL = {
+    vx4s: { id: 'vx4s', label: 'VX4S', brightness: exports.CHOICES_BRIGHTNESS, inputs: exports.CHOICES_INPUTS_VX4S, displayModes: exports.CHOICES_DISPLAYMODE,     presets: exports.CHOICES_PRESETS_VX4S ,piponoffs :exports.CHOICES_PIP_ONOFF },
+    
+    vx6s: { id: 'vx6s',label: 'VX6s', brightness: exports.CHOICES_BRIGHTNESS, inputs: exports.CHOICES_INPUTS_VX6S, displayModes: exports.CHOICES_DISPLAYMODE_VX6S, presets: exports.CHOICES_PRESETS_VX6S, workingModes: exports.CHOICES_WORKINGMODE_VX6S },
+    
+    novaProHD: { id: 'novaProHD', label: 'NovaPro HD', brightness: exports.CHOICES_BRIGHTNESS, inputs: exports.CHOICES_INPUTS_NOVAPROHD, displayModes: exports.CHOICES_DISPLAYMODE, piponoffs :exports.CHOICES_PIP_ONOFF },
+    
+    novaProUHDJr:{ id: 'novaProUHDJr', label: 'NovaPro UHD Jr', brightness: exports.CHOICES_BRIGHTNESS_VX1000, displayModes: exports.CHOICES_DISPLAYMODE_VX1000, presets: exports.CHOICES_PRESETS_NovaProUHDJR },
+    
+    mctrl4k:{ id: 'mctrl4k', label: 'MCTRL4K',   brightness: exports.CHOICES_BRIGHTNESS, inputs: exports.CHOICES_INPUTS_MCTRL4K, displayModes: exports.CHOICES_DISPLAYMODE},
+    
+    vx1000: { id: 'vx1000', label: 'VX1000', brightness: exports.CHOICES_BRIGHTNESS_VX1000, displayModes: exports.CHOICES_DISPLAYMODE_VX1000, inputs: exports.CHOICES_INPUTS_VX1000, presets: exports.CHOICES_PRESETS_VX1000, pipOnOffs: exports.CHOICES_PIP_ONOFF, 
+    layers: exports.CHOICES_LAYERS_VX1000, cardNo: exports.CHOICES_CARDNO_VX1000, layerPriority: exports.CHOICES_LAYERPRIORITY_VX1000, connectorCode: exports.CHOICES_CONNECTORCODE_VX1000},
+    
+    vx600: { id: 'vx600', label: 'VX600', brightness: exports.CHOICES_BRIGHTNESS_VX1000, displayModes: exports.CHOICES_DISPLAYMODE_VX1000, inputs: exports.CHOICES_INPUTS_VX600, presets: exports.CHOICES_PRESETS_VX1000 },
+    
+    vx16s: { id: 'vx16s', label: 'VX16S', brightness: exports.CHOICES_BRIGHTNESS , displayModes: exports.CHOICES_DISPLAYMODE_VX1000, presets: exports.CHOICES_PRESETS_VX1000 },
+    
+    j6: { id: 'j6', label: 'J6', displayModes: exports.CHOICES_DISPLAYMODE_VX1000, presets: exports.CHOICES_PRESETS_VX1000, workingModes: exports.CHOICES_WORKINGMODE_J6 }
+
+};
